+++ src/main/java/Yoj/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Yoj.command;
+package Yoj.tasks;
 public class Deadline extends Task {
     protected String by;
