+++ src/main/java/Yoj/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package Yoj.command;
+package Yoj.tasks;
 public class ToDo extends Task {
     public ToDo (String description){
         super(description);
